+++ src/types/StackNavigationType.ts
@@ -40,9 +40,6 @@ export interface DetailStackParamList extends ParamListBase {
   MarketReview: {
     marketId: number;
   };
-  CustomerReview: {
-    memberId: number;
-  };
 }
 
 export interface FeedStackParamList extends ParamListBase {
@@ -66,6 +63,9 @@ export interface MyPageStackParamList extends ParamListBase {
   Setting: undefined;
   Notice: undefined;
   Policy: undefined;
+  CustomerReview: {
+    memberId: number;
+  };
 }
 
 export interface RootStackParamList extends ParamListBase {
